requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.0
python-dotenv==0.20.0
pymongo==4.6.3
pydantic==2.4.0
marshmallow==3.19.0
mongoengine~=0.29.1
WTForms~=3.2.1
PyJWT~=2.10.1
Werkzeug~=3.1.3